Split the 200s function file, and stop 123 agent tests waiting on an image
Backend E2E's wall clock was one file. `test_function_e2e.py` was 200.4s of
the sandbox shard's 310s in 29 tests, and nothing in the layout could move
it: bin-packing the sandbox group four ways still returns a 200s bin, and
`--dist loadscope` puts a module's functions on one xdist worker, so no
worker count splits it either. Predicted wall goes 444s -> ~290s.

The file is now two, cut on what a test is *about* rather than on line
count -- which turns out to be the same cut, because 26 of 29 pay a cold
`docker run` and the item cost is nearly uniform. `test_function_e2e.py`
keeps the shape surface: lifecycle, the authorization matrix, record-grant
and table-facade writes. `test_function_execution_e2e.py` takes the runtime:
the queue, concurrency, timeouts, cancellation, connector operations, and
the four use-case sagas with no HTTP entry point of their own. 15 items and
15 (14 in the PR lane; the `slow` long-run test is scheduled-lane only).

Their shared helpers move to `app/modules/test_support/e2e/function_helpers.py`
and lose the leading underscore on the way, because a `_private` name
imported from another module is a lie about its scope. Not a sibling
`helpers.py`: `app/modules/function/tests/e2e/` has no `__init__.py` and the
suite runs under `--import-mode=importlib`, so a sibling is not reliably
importable, while that package already is and is where `waiters.py`,
`builders.py` and `e2e_authz.py` live.

Both halves carry `pytestmark = [e2e, workspace]`. That is load-bearing
rather than cosmetic: `workspace` is what the sandbox lane's filter selects
on, so a half that lost it would be collected, deselected, and reported as
a green shard that ran none of it -- `--verify` would stay happy and
`needs_sandbox_images` would stay "correct" because nothing ran. The
contract test added with the planner change fails if the two ever disagree.

Separately, the `agent` shard paid ~67s per run to restore the sandbox
image so that five of its 123 tests could provision a real workspace. Those
five live in two files, which now form their own `agent-workspace` group.
`agent` drops to `needs_sandbox_images: false` -- and drops it *derived*,
not declared: the contract test reads the answer off the sources, so
removing the last `fast_workspace` test from the shard is what changes the
flag. 327s -> ~206s, and the image lane absorbs the 99s.

That group is not folded into `sandbox`, because its marker filter is the
point. `fast_workspace` suppresses the auto-added `workspace` marker, so
these files select 27 tests under FAST_MARKERS and 38 under
SANDBOX_MARKERS -- the extra eleven are protected-lane `workspace` tests
that have never run on a PR, and moving them into the required lane is not
what this change is for.

Seven shards, not eight. An eighth bin would take the image lane from ~290s
to ~250s, but the packed lane is at ~288s and would become the wall, so it
buys nothing until the per-test fixture cost comes down. It would cost a
concurrency slot against a twenty-job cap that a full PR push already
exceeds.

Verified by collection rather than by reading: 819 selected node IDs, the
same 819 test names main selects, exactly 14 of them now reporting a
different file, and all 21 pairwise shard intersections empty. The function
directory runs green locally after the split -- 39 passed, the same 39.
